change meaning of DGglobals->fv_divf_extrap_s1
It was the parindex of fv_divf_extrap_s1. Now it is just the value of par
fv_divf_extrap_s1.

+++ src/main/dg/nmesh_dg.h
@@ -35,9 +35,7 @@ typedef struct {
   int fv_divf_adds_surface_fluxes; //if 1 fv_divf adds surf. fluxes
   int fv_surface_interp_mode; /* surf. interp. mode from fv_surface_interp */
   double fv_WENOm3_optw[2];   /* optimal weights for WENOm3 */
-  /* frequently used par indices */
-  //double fv_divf_extrap_s1;   /* value of fv_divf_extrap_s1 */
-  int fv_divf_extrap_s1;      /* index of par fv_divf_extrap_s1 */
+  double fv_divf_extrap_s1;   /* value of par fv_divf_extrap_s1 */
   int fv_divf_use_only_right_flux; //val of par fv_divf_use_only_right_flux
 } tDGglobals;
